requirements.txt
Flask
Flask-Mail
torch
torchvision
numpy
opencv-python
pillow
matplotlib
fpdf
python-dotenv
